Don't apply CTT fixes if RP-0 is loaded.
RP-0, a career mode for the RealismOverhaul universe, uses the CTT, but
sets its own tech positions. While it would be great if we could just do
an `:AFTER[DMagic]` to do this, we can't do that on MM stanzas that
don't have a `:FOR[DMagic]` attached to them. Because MM doens't yet
play nicely with having both `:FOR` and `:AFTER` tags, simplying adding
a `:FOR` doens't guarantee everyone gets the behaviour they want; plus
we can't exactly have `:FOR[DMagic]:AFTER[DMagic]`.

Consequetly, I've adjusted `DMagicCommunityTechTree.cfg` to only apply its
changes if RP-0 is not loaded. If you prefer we could instead
search for a special token, rather than RP-0 specifically. For example:
`:NEEDS[!CTT_Config]`. RP-0 and other career mods can simply
define `CTT_Config` if they wish to use the CTT, but do their own part
placement inside it.