Universal Storage parts will continue to work with missing animations.
It looks like the parts for universal storage are lacking a couple of animations,
which results in them being non functional because some curcial setup code is not
being executed on them.

I replaced the animation initialization with a common method that logs wa warning
to the log file if a referenced animation could not be found. This doesn't fix
missing animations, but it fixes the problems with Universal Storage... for the
most part. At least I'm able to use most experiments now.

Keep up the good work :)